packages/bsvibe-cli-base/pyproject.toml
[project]
name = "bsvibe-cli-base"
version = "0.1.0"
description = "BSVibe shared CLI foundation — Typer factory, profiles, keyring, OutputFormatter, device flow"
requires-python = ">=3.11"
dependencies = [
    "bsvibe-core",
    "bsvibe-authz",
    "pydantic>=2",
    "pydantic-settings>=2.5",
    "structlog>=24",
    "typer>=0.12",
    "httpx>=0.27",
    "keyring>=24",
    "pyyaml>=6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bsvibe_cli_base"]
